CORE-7637: moved the status endpoint implementation to a separate package
