+++ golang/src/permissions/restapi/impl/status/get_status.go
@@ -1,4 +1,4 @@
-package impl
+package status
 
 import (
 	"encoding/json"
